+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TC.java
@@ -6,38 +6,14 @@ import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 
 public class TC {
-    public static Trajectory RIGHT_startToCyclePole(SampleMecanumDrive drive, Pose2d pos) {
+    public static Trajectory RIGHT_deliverPreloadLeft(SampleMecanumDrive drive, Pose2d pos) {
         return drive.trajectoryBuilder(pos)
-                .forward(60)
+                .strafeLeft(36)
                 .build();
     }
-
-    public static Trajectory RIGHT_poleToStack(SampleMecanumDrive drive, Pose2d pos) {
+    public static Trajectory RIGHT_deliverPreloadForward(SampleMecanumDrive drive, Pose2d pos) {
         return drive.trajectoryBuilder(pos)
-                .forward(10)
-                .build();
-    }
-
-    public static Trajectory RIGHT_stackToPole(SampleMecanumDrive drive, Pose2d pos) {
-        // TODO: Write trajectory
-        return null;
-    }
-
-    public static Trajectory SCANPARK_forward(SampleMecanumDrive drive, Pose2d pos) {
-        return drive.trajectoryBuilder(pos)
-                .forward(24)
-                .build();
-    }
-
-    public static Trajectory SCANPARK_strafeLeft(SampleMecanumDrive drive, Pose2d pos) {
-        return drive.trajectoryBuilder(pos)
-                .strafeLeft(24)
-                .build();
-    }
-
-    public static Trajectory SCANPARK_strafeRight(SampleMecanumDrive drive, Pose2d pos) {
-        return drive.trajectoryBuilder(pos)
-                .strafeRight(24)
+                .forward(13)
                 .build();
     }
 }
