+++ fun743.py
@@ -1,0 +1,12 @@
+
+
+
+
+
+import webbrowser as wb
+import time
+while (True):
+	time.sleep(32)
+wb.open('https://youtu.be/dQw4w9WgXcQ')
+time.sleep(32)
+
